slicer: add function path to goto-instrument
This commit adds the command line argument handling for the
function path slicer.

To keep all possible exeuction traces that hit both function a and b, an
example invocation could look as follows:

goto-instrument --fp-reachability-slice a:b input output

Note, any trace that touches only a, but not b, will not be kept in the
program.