test: add fp-reachability test 2
This test tests whether paths are cut off in case a combination of
functions is specified. Function b is not on a path that executes
function c from the entry point main, hence, function b should
not be called. As c is on a single path when also executing
function a, both a and c should stay.